travel-planner/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
zhipuai==2.0.1
python-dotenv==1.0.0
httpx==0.25.2
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pandas==2.1.0
numpy==1.26.2
openpyxl==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
